requirements.txt
certifi==2020.6.20
chardet==3.0.4
cx-Oracle==8.0.1
idna==2.10
kafka-python==2.0.1
pkg-resources==0.0.0
requests==2.24.0
urllib3==1.25.10
